Move some tests into WarcraftClientTestsErrorConditions
Move some unit tests into a new WarcraftClientTestsErrorConditions class.
